fix: change attribute types in dtos

+++ src/main/java/com/es2/microservicos/dtos/responses/AluguelResponse.java
@@ -8,7 +8,7 @@ public record AluguelResponse(
         Long trancaInicio,
         LocalDateTime horaFim,
         Long trancaFim,
-        Integer cobranca,
+        double cobranca,
         Long ciclistaId
 ) {
 }
